requirements.txt
websockets
feedparser
httpx
python-dotenv
requests # SDK Dependency
websocket-client # SDK Dependency